+++ src/lib/auth-header.ts
@@ -1,26 +1,26 @@
 import jwt from "jsonwebtoken"
-import serverConfigs from "../../config/server"
+import serverConfigs from "../config"
 
 const cert = serverConfigs.jwtSecretKey
 class AuthHeader {
-	encodeUserLoginAuth(userId) {
-		return jwt.sign({ userId }, cert)
-	}
+  encodeUserLoginAuth(userId) {
+    return jwt.sign({ userId }, cert)
+  }
 
-	decodeUserLoginAuth(token) {
-		try {
-			return jwt.verify(token, cert)
-		} catch (error) {
-			return error
-		}
-	}
+  decodeUserLoginAuth(token) {
+    try {
+      return jwt.verify(token, cert)
+    } catch (error) {
+      return error
+    }
+  }
 
-	encodeUserPassword(token) {
-		return jwt.sign({ password: token }, cert)
-	}
+  encodeUserPassword(token) {
+    return jwt.sign({ password: token }, cert)
+  }
 
-	decodeUserPassword(token) {
-		return jwt.verify(token, cert)
-	}
+  decodeUserPassword(token) {
+    return jwt.verify(token, cert)
+  }
 }
 export default new AuthHeader()
